+++ OVSV1.xml
@@ -6,7 +6,7 @@
             <package name="org.dcsa.api_validator.ovs.v1"/>
             <package name="org.dcsa.api_validator.ovs.v1.schedules"/>
             <package name="org.dcsa.api_validator.ovs.v1.transportcalls"/>
-            <package name="org.dcsa.api_validator.ovs.v1.portcallevents"/>
+            <package name="org.dcsa.api_validator.ovs.v1.operationsevents"/>
             <package name="org.dcsa.api_validator.conf"/>
         </packages>
     </test> <!-- Test -->
