+++ src/test/java/org/dcsa/api_validator/ovs/v1/operationsevents/GetOperationsEventsTest.java
@@ -1,4 +1,4 @@
-package org.dcsa.api_validator.ovs.v1.portcallevents;
+package org.dcsa.api_validator.ovs.v1.operationsevents;
 
 import com.github.fge.jsonschema.cfg.ValidationConfiguration;
 import com.github.fge.jsonschema.main.JsonSchemaFactory;
@@ -9,7 +9,7 @@ import static com.github.fge.jsonschema.SchemaVersion.DRAFTV4;
 import static io.restassured.RestAssured.given;
 import static io.restassured.module.jsv.JsonSchemaValidator.matchesJsonSchemaInClasspath;
 
-public class GetPortCallEventsTest {
+public class GetOperationsEventsTest {
 
     JsonSchemaFactory jsonSchemaFactory = JsonSchemaFactory.newBuilder().setValidationConfiguration(ValidationConfiguration.newBuilder().setDefaultVersion(DRAFTV4).freeze()).freeze();
